# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TASK MANAGER API",
        "version": "1.0"
    },
    "paths": {
        "/api/app/user/auth/registration": {
            "post": {
                "tags": [
                    "User API"
                ],
                "summary": "Register user",
                "operationId": "6913dc25aaf5bf4aee19e7bb360c2c33",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "role": {
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "Abdullajon"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registration user and to login.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": {
                                                "username": "Abdullajon",
                                                "password": "$2y$10$sZZUV7ykVyAXT7/eDzGyReThtdkaLUK9jknrDEDNaeK71uKeQkfiO",
                                                "token": "ms7ST2opzINdvWCxLanr8zg21kOKFS",
                                                "updated_at": "2023-08-28T11:58:12.000000Z",
                                                "created_at": "2023-08-28T11:58:12.000000Z",
                                                "id": 1
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": "Login!"
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/app/user/auth/login": {
            "post": {
                "tags": [
                    "User API"
                ],
                "summary": "Login user",
                "operationId": "5b146fb05587bdf189f8c88b5b2b32ca",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "Abdullajon"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Check user **username** & **password** and to login.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": {
                                                "id": 1,
                                                "username": "Abdullajon",
                                                "password": "$2y$10$sZZUV7ykVyAXT7/eDzGyReThtdkaLUK9jknrDEDNaeK71uKeQkfiO",
                                                "token": "1XkHdP6TEgxACHCJlGsn4EitiPTIVB",
                                                "created_at": "2023-08-28T11:58:12.000000Z",
                                                "updated_at": "2023-08-28T12:03:25.000000Z"
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": "Login!"
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/app/user/get": {
            "get": {
                "tags": [
                    "User API"
                ],
                "summary": "Get user",
                "operationId": "78c77d22f61c38495e8485842b8fdd18",
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "description": "User token",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Check user **token** and get this user.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": {
                                                "id": 1,
                                                "username": "Abdullajon",
                                                "password": "$2y$10$sZZUV7ykVyAXT7/eDzGyReThtdkaLUK9jknrDEDNaeK71uKeQkfiO",
                                                "token": "1XkHdP6TEgxACHCJlGsn4EitiPTIVB",
                                                "created_at": "2023-08-28T11:58:12.000000Z",
                                                "updated_at": "2023-08-28T12:03:25.000000Z"
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": ""
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/app/user/list": {
            "get": {
                "tags": [
                    "User API"
                ],
                "summary": "Get user list",
                "operationId": "4828b4cee063f8c565fbcb36de7b877e",
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "description": "User token",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Check user **token** and get user list.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": [
                                                {
                                                    "id": 2,
                                                    "username": "Ali",
                                                    "password": "$2y$10$wnBgNzXvguR0QSAWJ9AehOnkKr1xzHTREEYOngzoUxZb0QC0cPQai",
                                                    "token": "8stomirSGdmP4TJPcML1gYBtAFsuYC",
                                                    "created_at": "2023-08-29T04:00:55.000000Z",
                                                    "updated_at": "2023-08-29T04:00:55.000000Z"
                                                },
                                                {
                                                    "id": 3,
                                                    "username": "Vali",
                                                    "password": "$2y$10$2gNZL7TnXp4RgcCWvGzjn.EgaOcAWrZgaLgi662ylRUVuA/CsYj1q",
                                                    "token": "QG45GsPO2CyMalWJjE2Cm2UPEpA0Yp",
                                                    "created_at": "2023-08-29T04:01:23.000000Z",
                                                    "updated_at": "2023-08-29T04:01:23.000000Z"
                                                }
                                            ]
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": ""
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/app/user/auth/logout": {
            "get": {
                "tags": [
                    "User API"
                ],
                "summary": "Logout user",
                "operationId": "454874fa08aa84bd6e12f074ae1b2e91",
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "description": "User token",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Check user **token** and logout this user.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": null
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": "Logout!"
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}